% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_main.R
\name{assign.breaks}
\alias{assign.breaks}
\title{Assign Breaks}
\usage{
assign.breaks(weekinquestion, break.type)
}
\arguments{
\item{weekinquestion}{The week to determine the last full break to include.
Must be in YYYY-MM-DD format}

\item{break.type}{The type of breaks to use. Options are seasonal or monthly}
}
\description{
Identify break points according to day of year, and drop any that exceed
the weekinquestion variable (if you want the entire year, select December 33)
Does not correct for leap years, but these will be corrected for in later steps,
such as in the convert.env.data function (correcting here made it difficult
to correct appropriately in other parts of the code)
}
