% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_main.R
\name{calculate.MLE.v2}
\alias{calculate.MLE.v2}
\title{Calculate MLE for a specified subset of the data using R}
\usage{
calculate.MLE.v2(md, temporal.resolution = "annual")
}
\arguments{
\item{md}{Data on mosquito pools tested for the disease. Must be formatted with 4 columns: location (the spatial unit, e.g. county), col_date: the date the mosquitoes were tested, wnv_result: whether or not the pool was positive, pool_size: the number of mosquitoes tested in the pool. A fifth column species is optional but is not used by the code}

\item{temporal.resolution}{Must be set to 'annual' for now. Ideally support for finer-scale resolution will be added.}
}
\value{
Mosquito data compiled to give estimated mosquito infection rates by spatial unit and temporal resolution
}
\description{
Modified from wnv_hlpr.R
}
