% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_main.R
\name{convert.env.data}
\alias{convert.env.data}
\title{Convert environmental data to the seasonal format used by the RF1 model}
\usage{
convert.env.data(weather.data, all.locations, season.breaks,
  leap.correct = 1)
}
\arguments{
\item{weather.data}{Data on weather variables to be included in the analysis.
See the read.weather.data function for details about data format. The
read.weather.data function from ArboMAP is a useful way to process one or
more data files downloaded via Google Earth Engine.}

\item{all.locations}{A list of counties included in the analysis (this
is for ensuring that a county and year that does not have a human case is
treated as a 0)}

\item{season.breaks}{The break points to use for calculating summaries. These should be in the format of day of year.
For example, c(31, 59) would create monthly summaries for January and February. the assign.breaks function
can assist with generating the breakpoints automatically in the correct format.}

\item{leap.correct}{0: no leap year corrections will be made, all years will have the same day of year.
Day 366 may be lost unless the last break is >366. 1: leap year corrections will be made for years with leap years}
}
\description{
Convert environmental data to the seasonal format used by the RF1 model
}
