% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_simulated_data.R
\docType{data}
\name{human.data}
\alias{human.data}
\title{Human Data}
\format{An object of class \code{data.frame} with 11 rows and 2 columns.}
\source{
See file rf1_simulated_data.R
}
\usage{
human.data
}
\description{
Simulated unrealistic human case data to test the Random
Forest model. All cases occur on the same date of the year.
Five years each with one case per district are simulated,
then approximately half the cases were dropped at random.
}
\keyword{datasets}
