% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_simulated_data.R
\docType{data}
\name{human.results}
\alias{human.results}
\title{Human Results}
\format{An object of class \code{list} of length 6.}
\usage{
human.results
}
\description{
Results from the Random Forest model using the simulated human data.
Used in the unit tests to separate running the models from 
extracting the the model outputs
}
\keyword{datasets}
