% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_main.R
\name{is.leap}
\alias{is.leap}
\title{Function to identify is a leap year or not}
\usage{
is.leap(year)
}
\arguments{
\item{year}{The year to check if it is a leap-year or not}
}
\value{
1 for leap year, 0 for non-leap year
}
\description{
Takes a year, returns a 1 if it is a leap year, based on the result of the get.days function
}
