% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_simulated_data.R
\docType{data}
\name{mosq.data}
\alias{mosq.data}
\title{Mosquito Data}
\format{An object of class \code{data.frame} with 40 rows and 5 columns.}
\source{
See file rf1_simulated_data.R
}
\usage{
mosq.data
}
\description{
Simulated mosquito data for mosquito pools. 
Assumed two trap-nights per year per county.
Positive pools were determined at random, with values drawn
from a uniform random distribution between 0 and 0.7.
Values below 0.5 were rounded to 0, values above were rounded
to 1.
}
\keyword{datasets}
