% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_simulated_data.R
\docType{data}
\name{mosquito.results}
\alias{mosquito.results}
\title{Mosquito Results}
\format{An object of class \code{list} of length 6.}
\usage{
mosquito.results
}
\description{
Results from using the simulated mosquito data generated by the 
Random Forest model. Used in the unit tests to separate testing
of the model from testing extraction of the model outputs
}
\keyword{datasets}
