% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf1_simulated_data.R
\docType{data}
\name{weather.data}
\alias{weather.data}
\title{Weather Data}
\format{An object of class \code{data.frame} with 7304 rows and 10 columns.}
\source{
See file rf1_simulated_data.R
}
\usage{
weather.data
}
\description{
Data follows ArboMAP format (except missing date and districtdate fields)
Data are entirely simulated to represent anomalies drawn from a
normal distribution with mean of 0 and standard deviation of 1.
They have no actual relationship with the human cases or mosquito cases.
}
\keyword{datasets}
